stock-news-service/pyproject.toml
[project]
name = "stock-news-service"
version = "0.1.0"
description = "Stock news microservice with FastAPI, SQLAlchemy, and Finnhub"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.2.2",
    "finnhub-python>=2.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]

[tool.uv]
package = false
